refactor!: simplify hooks namespace
