+++ open_feature/hooks/hook_support.py
@@ -5,9 +5,7 @@ from functools import reduce
 from open_feature.evaluation_context import EvaluationContext
 from open_feature.flag_evaluation.flag_evaluation_details import FlagEvaluationDetails
 from open_feature.flag_evaluation.flag_type import FlagType
-from open_feature.hooks.hook import Hook
-from open_feature.hooks.hook_context import HookContext
-from open_feature.hooks.hook_type import HookType
+from open_feature.hooks import Hook, HookContext, HookType
 
 
 def error_hooks(
