--- open_feature/hooks/hook_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class HookType(Enum):
-    BEFORE = "before"
-    AFTER = "after"
-    FINALLY_AFTER = "finally_after"
-    ERROR = "error"
