+++ open_feature/open_feature_api.py
@@ -2,7 +2,7 @@ import typing
 
 from open_feature.evaluation_context import EvaluationContext
 from open_feature.exception import GeneralError
-from open_feature.hooks.hook import Hook
+from open_feature.hooks import Hook
 from open_feature.open_feature_client import OpenFeatureClient
 from open_feature.provider.metadata import Metadata
 from open_feature.provider.no_op_provider import NoOpProvider
